Revert "Replace float with numeric"
This reverts commit 67470bc408330a927991341e3abb46513a47b4f6.

+++ lib/budget/accounts/account.ex
@@ -6,7 +6,7 @@ defmodule Budget.Accounts.Account do
 
   schema "accounts" do
     field(:name, :string)
-    field(:balance, :decimal)
+    field(:balance, :float)
     field(:debt, :boolean)
     timestamps()
